Refactor and updating the plugin for Net2Plan 0.4.0 INCOMPLETE
