+++ src/main/java/es/upct/girtel/net2plan/plugins/activepce/utils/GenericServer.java
@@ -8,7 +8,7 @@
  *  Contributors:
  */
 
-package es.upct.girtel.net2plan.plugins.abnopce.utils;
+package es.upct.girtel.net2plan.plugins.activepce.utils;
 
 import com.net2plan.interfaces.networkDesign.Net2PlanException;
 import java.io.IOException;
