+++ src/main/java/es/upct/girtel/net2plan/plugins/activepce/utils/LoadBalancing_UPCT.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package es.upct.girtel.net2plan.plugins.abnopce.utils;
+package es.upct.girtel.net2plan.plugins.activepce.utils;
 
 import es.tid.pce.pcep.objects.LoadBalancing;
