+++ src/main/java/es/upct/girtel/net2plan/plugins/activepce/utils/RequestHandler.java
@@ -8,7 +8,7 @@
  *  Contributors:
  */
 
-package es.upct.girtel.net2plan.plugins.abnopce.utils;
+package es.upct.girtel.net2plan.plugins.activepce.utils;
 
 import java.io.IOException;
 import java.io.InputStream;
